+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Fixed
 
 - More fixes and improvements regarding encoding/decoding. Deprecated functions escape() and unescape() are no longer being used.
+- Fixed an issue with the 'updateBoolean' query parameter not working properly on /thread/:id.
 
 ## [1.0.1] - 2023-03-09